+++ packages/core-api/lib/utils/generate-cache-key.js
@@ -1,0 +1,5 @@
+module.exports = value =>
+  require('crypto')
+    .createHash('sha256')
+    .update(JSON.stringify(value))
+    .digest('hex')