+++ packages/core-api/lib/versions/1/utils.js
@@ -25,6 +25,21 @@ const respondWith = (data, error = false) =>
   error ? { error: data, success: false } : { ...data, success: true }
 
 /**
+ * Respond with data from cache.
+ * @param  {Object} data
+ * @param  {Hapi.Toolkit} h
+ * @return {Object}
+ */
+const respondWithCache = (data, h) => {
+  const { value, cached } = data
+  const lastModified = cached ? new Date(cached.stored) : new Date()
+
+  return value.isBoom
+    ? h.response(value.output.payload).code(value.output.statusCode)
+    : h.response(value).header('Last-modified', lastModified.toUTCString())
+}
+
+/**
  * Transform the given data into a resource.
  * @param  {Hapi.Request} request
  * @param  {Object} data
@@ -49,6 +64,7 @@ const toCollection = transformCollection
 module.exports = {
   paginate,
   respondWith,
+  respondWithCache,
   toResource,
   toCollection,
 }