+++ packages/core-container/lib/registrars/plugin.js
@@ -18,7 +18,7 @@ module.exports = class PluginRegistrars {
     this.container = container
     this.plugins = this.__loadPlugins()
     this.resolvedPlugins = []
-    this.options = options
+    this.options = this.__castOptions(options)
     this.deregister = []
   }
 
@@ -153,6 +153,27 @@ module.exports = class PluginRegistrars {
     if (this.options.options && this.options.options[name]) {
       options = Hoek.applyToDefaults(options, this.options.options[name])
     }
+
+    return this.__castOptions(options)
+  }
+
+  /**
+   * When the env is used to overwrite options, we get strings even if we
+   * expect a number. This is in most cases not desired and leads to side-
+   * effects. Here is assumed all numeric strings except blacklisted ones
+   * should be treated as numbers.
+   * @param {Object} options
+   * @return {Object} options
+   */
+  __castOptions(options) {
+    const blacklist = []
+    const regex = new RegExp(/^\d+$/)
+    Object.keys(options).forEach(key => {
+      const value = options[key]
+      if (isString(value) && !blacklist.includes(key) && regex.test(value)) {
+        options[key] = +value
+      }
+    })
 
     return options
   }