+++ packages/core/lib/config/testnet.live/plugins.js
@@ -30,7 +30,6 @@ module.exports = {
     enabled: true,
     maxTransactionsPerSender:
       process.env.ARK_TRANSACTION_POOL_MAX_PER_SENDER || 300,
-    whitelist: [],
     allowedSenders: [],
   },
   '@arkecosystem/core-p2p': {