+++ packages/crypto/lib/utils/format-arktoshi.js
@@ -7,19 +7,10 @@ const configManager = require('../managers/config')
  * @return {String}
  */
 module.exports = amount => {
-  let localeString = (+amount / ARKTOSHI).toLocaleString('en', {
-    minimumFractionDigits: 8,
+  const localeString = (+amount / ARKTOSHI).toLocaleString('en', {
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 8,
   })
-
-  if (localeString.includes('.')) {
-    while (localeString.slice(-1) === '0') {
-      localeString = localeString.slice(0, -1)
-    }
-
-    if (localeString.slice(-1) === '.') {
-      localeString = localeString.slice(0, -1)
-    }
-  }
 
   return `${localeString} ${configManager.config.client.symbol}`
 }