+++ packages/core-transaction-pool/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.2.0 - 2018-11-28
+
 ### Added
 
 - Delete pool wallet if no ballance or no transactions in pool
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Handle dynamic fees
 - Better transaction ping
 - Broadcasting on launch of relay
+- Cap the number of transactions the pool can hold
 
 ### Changed
 
@@ -30,6 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Adjusted timeouts & lifetimes
 - No longer add transactions that do not meet the fee requirements to the pool (BTC/ETH behaviour)
 - No longer decline transactions that exceed the static fees
+- Simplified the broadcasting logic
+- Broadcast transactions when the pool is full
 
 ### Fixed
 
@@ -40,6 +45,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Properly determine valid transactions based on their type
 - Handle unexpected errors in the guard
 - Revert transactions with non-matching dynamic fees
+- Revert excess transactions
+- Reject transactions that have a timestamp from the future
+- Remove invalid transactions from pool when receiving a bad block
 
 ## 0.1.1 - 2018-06-14